// apps/frontend/src/api/model/pageBaseCreateType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Workspace Management API
 * API specification for managing users, workspaces, and pages.
 * OpenAPI spec version: 1.0.0
 */

/**
 * The type of the page, which determines its properties and behavior.
 */
export type PageBaseCreateType = typeof PageBaseCreateType[keyof typeof PageBaseCreateType];


export const PageBaseCreateType = {
  course: 'course',
  assignment: 'assignment',
  folder: 'folder',
  note: 'note',
} as const;