// apps/frontend/src/api/model/pagePropertiesAssignmentType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Workspace Management API
 * API specification for managing users, workspaces, and pages.
 * OpenAPI spec version: 1.0.0
 */

export type PagePropertiesAssignmentType = typeof PagePropertiesAssignmentType[keyof typeof PagePropertiesAssignmentType];


export const PagePropertiesAssignmentType = {
  assignment: 'assignment',
} as const;