// apps/frontend/src/api/model/pagePropertiesCourseType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Workspace Management API
 * API specification for managing users, workspaces, and pages.
 * OpenAPI spec version: 1.0.0
 */

export type PagePropertiesCourseType = typeof PagePropertiesCourseType[keyof typeof PagePropertiesCourseType];


export const PagePropertiesCourseType = {
  course: 'course',
} as const;