// apps/frontend/src/api/model/pagePropertiesNoteType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Workspace Management API
 * API specification for managing users, workspaces, and pages.
 * OpenAPI spec version: 1.0.0
 */

export type PagePropertiesNoteType = typeof PagePropertiesNoteType[keyof typeof PagePropertiesNoteType];


export const PagePropertiesNoteType = {
  note: 'note',
} as const;